First working version of displaying program output
The output of a simple program hardcoded into the MIPS CPU, in this case, 4 + 10, is displayed via the LED lights on the Arty A7. The LEDs represent a "ready" bit followed by bits 7-4 of the output, and the RGB LEDs represent bits 3-0 of the output. For simplicity, the RGBs are treated as binary (either 111 or 000), although in theory, this system could be used to support 16 bits of output.

Additionally, to facilitate testing and ease further development, the `arty.ml` file now just calls the `Wrapper` module, which resides in a new `hardcaml_mips_arty` library.