feat: add uart transmission support (#4)
Currently, our FPGA outputs results by lighting up some combination of the 8 available LEDs. We decided to treat the RGB LEDs as binary, but even if we hadn't, we'd still only have 16 bits of output. This is enough to demonstrate that our CPU can execute programs correctly on the hardware, but doesn't really model realistic I/O, especially since there's no input mechanism.

The next stage of our project is adding memory mapped I/O; in other words, having certain addresses in data memory that, when read to / written from, will use the UART bus (or LEDs, or something else) instead of just the RAM memory. This motivates a refactoring of our architecture to be more realistic, with the current `wrapper.ml` acting as a bus between the CPU and I/O devices (which will include the current data memory).

The first step towards this is supporting basic transmissions via UART. In this commit, we keep our current "output the writeback data once the program has concluded" strategy, but instead of lighting up LEDs, we send UART output. As part of this, we implemented a state-machine-based buffer that queues up words of data, and outputs bytes in the proper format at the correct baud.

Co-authored-by: Alexander Skvortsov <38059171+askvortsov1@users.noreply.github.com>