feat: refactor to use bus architecture (#5)
Read/write addresses, write data, and read/write enable flags are now outputs of the datapath. The read data is now an input. This allows us to move I/O handling logic outside of the datapath to a `cpu` wrapper component that interfaces with the Arty FPGA's hardware.

`writeback_data` and `writeback_pc` are still outputs so that integration tests for basic operations can still work. Ideally we could expose these signals to tests without making them parts of the input/output.

Co-authored-by: Mayur Deshpande <mayur0@outlook.com>