feat: define delayed test IO
This component is for testing, not as an actual production destination.

When a read/write is triggered, it stalls for a fixed number of cycles (2 for reads, 4 for writes) before outputting a fixed value in case of reads (arbitrarily 5).

This is used as a simple example of an async IO interface to test that the datapath pauses properly without needing to worry about the correctness  of UART as well.