fix: don't forward from first stall instruction
When an instruction depends on a preceding lw, we stall and repeat it twice. The first iteration will have incorrect inputs and be wrong, but the second one should be correct, since we can now forward from memory. Here, we eliminate forwards from the first (broken) instruction to the second instruction.